MacOSXS/OpenStep/NextStep compiler warned about comparision between signed and unsigned.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@976 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csnetdrv/sockets/netsdrv.cpp
@@ -218,7 +218,7 @@ STDMETHODIMP csNetworkDriverSockets::Receive(DWORD dwID, DWORD *lpdwBytesToRecei
 	}
 
 #if !defined(NO_SOCKETS_SUPPORT)
-	if((*lpdwBytesToReceive = recv(Socket[dwID], lpDataBuffer, sizeof(lpDataBuffer), 0)) != -1) return S_OK;
+	if((*lpdwBytesToReceive = recv(Socket[dwID], lpDataBuffer, sizeof(lpDataBuffer), 0)) != (DWORD)-1) return S_OK;
 #else
 	return S_OK;
 #endif
